Reworked Switches, Added FIFO, UART, and broken Joystick code
Notes: I renamed the ADC_Init in texas.c to ADC_Init2. They both seem to do the same thing so I'm not super worried it broke anything. We also don't use texas to debug anything so we should be fine. Currently the click stick does not work, and I haven't really started programming the potentiometers in them.