Color system, struct changes, timer system, game over screen
The top or bottom button now lets you swap colors. Currently enemies are not spawned in as a color, although they have it as an internal value. Collisions have not been update to check color matches as a result. The sprite struct now stores different states for a character as a 2d array, with the x axis being health and the y being color. A reworked game over screen has been added. Enemies can now be spawned with a new function. New enemy types are on the way soon! A timer system using a counter that updates every frame has been added, allowing for an actual level to be designed.